Split out some of the rendering system specific engine stuff into their own classes.
